Split CalcSpectrum to its own component
Closes #23
